build(blog): split blog client into `blog-client-api` and `blog-client-webmvc` and update gradle configs
- Replaced `blogClient` with `blogClientApi` and `blogClientWebMvc` in `gradle.properties`.
- Updated references in `build.gradle.kts` and `settings.gradle.kts` to new module coordinates.
- Rationale: clarify dependency boundaries and separate API contract from WebMVC adapter in hexagonal multi-module setup.
