+++ gradle.properties
@@ -35,7 +35,8 @@ blogReadModel=:blog:blog-api:blog-readmodel
 blogApplication=:blog:blog-application
 blogRdbAdapter=:blog:blog-rdb-adapter
 blogWebMvcAdapter=:blog:blog-webmvc
-blogClient=:blog:blog-client
+blogClientApi=:blog:blog-client-api
+blogClientWebMvc=:blog:blog-client-webmvc
 
 # series
 series=:series
