+++ services/blog/export/client-api/build.gradle.kts
@@ -1,0 +1,5 @@
+val blogApi: String by project
+
+dependencies {
+    api(project(blogApi))
+}
