+++ services/blog/export/client-api/src/main/java/nettee/blolet/blog/export/client/api/BlogClientDto.java
@@ -1,4 +1,4 @@
-package nettee.blolet.blog.export.client;
+package nettee.blolet.blog.export.client.api;
 
 import lombok.Builder;
